example using a explicit initial state
