+++ src/main/scala/com/lightbend/lagom/AppInitialState.scala
@@ -8,23 +8,23 @@ import scala.concurrent.Await
 import scala.concurrent.duration._
 import scala.concurrent.ExecutionContext.Implicits.global
 
-object App {
+object AppInitialState {
 
   def main(args: Array[String]): Unit = {
 
-    val entityRef = new PersistenceEntityRef(new AccountEntity)
+    val entityRef = new PersistenceEntityRef(new AccountEntityWithInitState)
 
     val result =
-    for {
-      _ <- entityRef ? Deposit(100)
-      _ <- entityRef ? Deposit(120)
-      _ <- entityRef ? Withdraw(80)
-      balance <- entityRef ? GetBalance
-      state <- entityRef ? GetState
-    } yield {
-      println("Balance is: " + balance)
-      println("State is: " + state)
-    }
+      for {
+        _ <- entityRef ? Deposit(100)
+        _ <- entityRef ? Deposit(120)
+        _ <- entityRef ? Withdraw(80)
+        balance <- entityRef ? GetBalance
+        state <- entityRef ? GetState
+      } yield {
+        println("Balance is: " + balance)
+        println("State is: " + state)
+      }
 
     Await.ready(result, 500.millis)
